+++ chatlearn/runtime/utils.py
@@ -15,7 +15,6 @@
 """runtime utils"""
 
 import ast
-import concurrent.futures
 import textwrap
 import inspect
 
@@ -74,16 +73,3 @@ class FlowParser:
         node_iter.visit(ast.parse(code))
         return self.model_to_call_func
 
-def execute_in_parallel(function, arguments):
-    if len(arguments) == 1:
-        return function(*arguments[0])
-    results = []
-    with concurrent.futures.ThreadPoolExecutor() as executor:
-        # Using list comprehension to handle the results
-        futures = [executor.submit(function, *args) for args in arguments]
-        for _future in concurrent.futures.as_completed(futures):
-            try:
-                results.append(_future.result())
-            except Exception as e:
-                print(f"Thread generated an exception: {e}")
-    return results
